+++ stdlib_cases/case_241_fs_mkdir_existing_throw.nc
@@ -4,12 +4,10 @@ import io
 # STDOUT: fs.mkdir failed
 fun main() {
     let dir = "__nc_case_241_dir"
-    if fs.exists(dir) { fs.remove(dir) }
-    fs.mkdir(dir)
-    try {
-        fs.mkdir(dir)
-    } catch e {
-        io.println(e)
+    if fs.exists(dir) { fs.remove(dir)!! }
+    fs.mkdir(dir)!!
+    if fs.mkdir(dir) is err {
+        io.println("fs.mkdir failed")
     }
-    fs.remove(dir)
+    fs.remove(dir)!!
 }
